+++ config/routes.rb
@@ -1,0 +1,3 @@
+get 'redmine_goodies_edit_field', :to => 'redmine_goodies_quick_edit#edit_field', :as => 'redmine_goodies_edit_field'
+post 'redmine_goodies_update_field', :to => 'redmine_goodies_quick_edit#update_field', :as => 'redmine_goodies_update_field'
+