+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Not released yet:
+* [Actions triggered when fields are changed](https://github.com/famiprog/redmine_goodies/pull/11)
+
 ## 1.1.3:
 * [Fixing uncaught ReferenceError: jsToolBar is not defined](https://github.com/famiprog/redmine_goodies/pull/12)